refactor(query): Split runner/explain/mod.rs into submodules
Split the 1,532-line explain module into 5 focused files:
- mod.rs (140 lines): Public entry points + module declarations
- aggregate.rs (245 lines): Top-level aggregate explain logic
- select.rs (270 lines): Select explain variants
- execute.rs (425 lines): Query execution with metrics
- mutation.rs (489 lines): Mutation explain paths

Purely mechanical move — no logic changes. All moved methods
changed from private to pub(crate) visibility to support split
impl blocks across modules.

Closes #329
